+++ lib/shims/bootpay_web.dart
@@ -2,22 +2,16 @@
 library bootpay_api;
 
 import 'dart:convert';
+import 'dart:js_interop';
 
 import 'package:bootpay/api/bootpay_analytics.dart';
 import 'package:bootpay/bootpay_widget_api.dart';
 import 'package:bootpay/model/stat_item.dart';
 import 'package:http/src/response.dart';
-import 'package:js/js.dart';
 import 'package:flutter/material.dart';
 import '../bootpay.dart';
 import '../bootpay_api.dart';
 import '../model/payload.dart';
-
-@JS('Promise')
-class Promise<T> {
-  external Promise (void executor(void resolve(T result), Function reject));
-  external Promise then (void onFulfilled(T result), [Function onRejected]);
-}
 
 @JS()
 external void _jsBeforeLoad();
@@ -33,32 +27,28 @@ external void _setLocale(String locale);
 external void _removePaymentWindow();
 
 @JS()
-external void _dismiss(BuildContext context);
-@JS()
 external void _transactionConfirm();
-@JS()
-external void _addCloseEvent();
 
 @JS('BootpayClose')
-external set _BootpayClose(void Function() f);
+external set _BootpayClose(JSFunction f);
 
 @JS('BootpayCancel')
-external set _BootpayCancel(void Function(String) f);
+external set _BootpayCancel(JSFunction f);
 
 @JS('BootpayDone')
-external set _BootpayDone(void Function(String) f);
+external set _BootpayDone(JSFunction f);
 
 @JS('BootpayIssued')
-external set _BootpayIssued(void Function(String) f);
+external set _BootpayIssued(JSFunction f);
 
 @JS('BootpayConfirm')
-external set _BootpayConfirm(bool Function(String) f);
+external set _BootpayConfirm(JSFunction f);
 
 @JS('BootpayAsyncConfirm')
-external set _BootpayAsyncConfirm(Promise Function(String) f);
+external set _BootpayAsyncConfirm(JSFunction f);
 
 @JS('BootpayError')
-external set _BootpayError(void Function(String) f);
+external set _BootpayError(JSFunction f);
 
 class BootpayPlatform extends BootpayApi with BootpayWidgetApi {
   BootpayDefaultCallback? _callbackCancel;
@@ -70,14 +60,22 @@ class BootpayPlatform extends BootpayApi with BootpayWidgetApi {
   BootpayDefaultCallback? _callbackDone;
 
   BootpayPlatform() {
-    _BootpayClose = allowInterop(onClose);
-    _BootpayCancel = allowInterop(onCancel);
-    _BootpayDone = allowInterop(onDone);
-    _BootpayIssued = allowInterop(onIssued);
-    _BootpayConfirm = allowInterop(onConfirm);
-    _BootpayAsyncConfirm = allowInterop(onConfirmAsync); //js에서 BootpayAsyncConfirm 호출시 onConfirmAsync 수행
-    _BootpayError = allowInterop(onError);
+    _BootpayClose = onClose.toJS;
+    _BootpayCancel = _onCancelWrapper.toJS;
+    _BootpayDone = _onDoneWrapper.toJS;
+    _BootpayIssued = _onIssuedWrapper.toJS;
+    _BootpayConfirm = _onConfirmWrapper.toJS;
+    _BootpayAsyncConfirm = _onConfirmAsyncWrapper.toJS;
+    _BootpayError = _onErrorWrapper.toJS;
   }
+
+  // JS에서 JSString으로 받아서 Dart String으로 변환하는 wrapper 함수들
+  void _onCancelWrapper(JSString data) => onCancel(data.toDart);
+  void _onDoneWrapper(JSString data) => onDone(data.toDart);
+  void _onIssuedWrapper(JSString data) => onIssued(data.toDart);
+  bool _onConfirmWrapper(JSString data) => onConfirm(data.toDart);
+  JSPromise<JSBoolean> _onConfirmAsyncWrapper(JSString data) => onConfirmAsync(data.toDart);
+  void _onErrorWrapper(JSString data) => onError(data.toDart);
 
 
   @override
@@ -254,17 +252,18 @@ class BootpayPlatform extends BootpayApi with BootpayWidgetApi {
     return false;
   }
 
-  Promise onConfirmAsync(String data)  {
-
-    return Promise<bool>(allowInterop((resolve, reject) async {
-      if(this._callbackAsyncConfirm != null) {
-
-        bool result = await this._callbackAsyncConfirm!(data);
-        resolve(result);
-      } else {
-        resolve(false);
+  JSPromise<JSBoolean> onConfirmAsync(String data) {
+    return JSPromise<JSBoolean>((JSFunction resolve, JSFunction reject) {
+      Future<void> doAsync() async {
+        if (_callbackAsyncConfirm != null) {
+          bool result = await _callbackAsyncConfirm!(data);
+          resolve.callAsFunction(null, result.toJS);
+        } else {
+          resolve.callAsFunction(null, false.toJS);
+        }
       }
-    }));
+      doAsync();
+    }.toJS);
   }
 
   void onDone(String data) {